some stat, io gaps

+++ shedskin/lib/os/path.cpp
@@ -263,7 +263,7 @@ double getmtime(str *filename) {
     Return the last modification time of a file, reported by os.stat().
     */
 
-    return (double)(__os__::stat(filename))->__ss_st_mtime;
+    return (__os__::stat(filename))->__ss_st_mtime;
 }
 
 double getatime(str *filename) {
@@ -271,7 +271,7 @@ double getatime(str *filename) {
     Return the last access time of a file, reported by os.stat().
     */
 
-    return (double)(__os__::stat(filename))->__ss_st_atime;
+    return (__os__::stat(filename))->__ss_st_atime;
 }
 
 double getctime(str *filename) {
@@ -279,7 +279,7 @@ double getctime(str *filename) {
     Return the metadata change time of a file, reported by os.stat().
     */
 
-    return (double)(__os__::stat(filename))->__ss_st_ctime;
+    return (__os__::stat(filename))->__ss_st_ctime;
 }
 
 __ss_bool islink(str *path) {
@@ -970,7 +970,7 @@ double getmtime(str *filename) {
     Return the last modification time of a file, reported by os.stat()
     */
 
-    return (double)(__os__::stat(filename))->__ss_st_mtime;
+    return (__os__::stat(filename))->__ss_st_mtime;
 }
 
 double getatime(str *filename) {
@@ -978,7 +978,7 @@ double getatime(str *filename) {
     Return the last access time of a file, reported by os.stat()
     */
 
-    return (double)(__os__::stat(filename))->__ss_st_atime;
+    return (__os__::stat(filename))->__ss_st_atime;
 }
 
 double getctime(str *filename) {
@@ -986,7 +986,7 @@ double getctime(str *filename) {
     Return the creation time of a file, reported by os.stat().
     */
 
-    return (double)(__os__::stat(filename))->__ss_st_ctime;
+    return (__os__::stat(filename))->__ss_st_ctime;
 }
 
 __ss_bool islink(str *path) {
